+++ macos/Classes/CameraMacosPlugin.swift
@@ -6,12 +6,11 @@ import CoreVideo.CVPixelBuffer
 import Accelerate.vecLib
 import CoreImage.CIContext
 
-public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate, AVCaptureAudioDataOutputSampleBufferDelegate, AVAssetWriterDelegate, AVCaptureFileOutputRecordingDelegate,
-    FlutterStreamHandler{
+public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate, AVCaptureAudioDataOutputSampleBufferDelegate, AVAssetWriterDelegate, AVCaptureFileOutputRecordingDelegate {
     
     let registry: FlutterTextureRegistry
     let outputChannel: FlutterMethodChannel!
-    var sink: FlutterEventSink!
+    var imageStreamHandler: ImageStreamHandler!
     
     // Texture id of the camera preview
     var textureId: Int64!
@@ -69,22 +68,13 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
     var zoomPixelBuffer: CVImageBuffer?
     
     var orientation:CGFloat = 0
+
+    var isVideoMirrored: Bool = true
     
     init(_ registry: FlutterTextureRegistry, _ outputChannel: FlutterMethodChannel) {
         self.registry = registry
         self.outputChannel = outputChannel
         super.init()
-    }
-    // FlutterStreamHandler
-    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
-        sink = events
-        return nil
-    }
-    
-    // FlutterStreamHandler
-    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
-        sink = nil
-        return nil
     }
     
     public static func register(with registrar: FlutterPluginRegistrar) {
@@ -97,29 +87,33 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
         instance.factory = factory
         
         // Channel for communicating with platform plugins using event streams
-        let event = FlutterEventChannel(name:"camera_macos/stream", binaryMessenger: registrar.messenger)
-        registrar.addMethodCallDelegate(instance, channel: outputChannel)
-        event.setStreamHandler(instance)
+        instance.imageStreamHandler = ImageStreamHandler(id: "camera_macos/stream", messenger: registrar.messenger)
     }
     
     public func copyPixelBuffer() -> Unmanaged<CVPixelBuffer>? {
         if latestBuffer == nil {
             return nil
         }
-        if self.sink != nil{
+        if let imageStreamHandler = self.imageStreamHandler {
             let u = imageFromSampleBuffer(imageBuffer: latestBuffer)!
             let bytesPerRow = u.bytesPerRow
             let width = Int(u.size.width)
             let height = Int(u.size.height)
             
-            let newData:Data = Data(bytes: u.bitmapData!, count: Int(bytesPerRow*height))
-            
-            self.sink([
-                "width": width,
-                "height": height,
-                "bytesPerRow": bytesPerRow,
-                "data": newData,
-            ] as [String:Any]);
+            DispatchQueue.main.async {
+                do {
+                    let newData:Data = Data(bytes: u.bitmapData!, count: Int(bytesPerRow*height))
+                    
+                    try imageStreamHandler.success([
+                        "width": width,
+                        "height": height,
+                        "bytesPerRow": bytesPerRow,
+                        "data": newData,
+                    ] as [String:Any]);
+                } catch(let err) {
+                    imageStreamHandler.error(code: "IMAGE_STREAM_ERROR", message: err.localizedDescription)
+                }
+            }
         }
         
         return Unmanaged<CVPixelBuffer>.passRetained(latestBuffer)
@@ -153,6 +147,9 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
         case "setOrientation":
             let arguments = call.arguments as? Dictionary<String, Any> ?? [:]
             orientation = arguments["orientation"] as? Double ?? 0
+        case "setVideoMirrored":
+            let arguments = call.arguments as? Dictionary<String, Any> ?? [:]
+            isVideoMirrored = arguments["isVideoMirrored"] as? Bool ?? true
         case "destroy":
             destroy(result)
         case "setFocusPoint":
@@ -301,6 +298,8 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                 } else {
                     newCameraObject = capturedVideoDevices.first
                 }
+
+                self.isVideoMirrored = arguments["isVideoMirrored"] as? Bool ?? true
                 
                 self.orientation = arguments["orientation"] as? Double ?? 0
                 
@@ -500,7 +499,7 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                             self.captureSession.addOutput(videoOutput)
                             for connection in videoOutput.connections {
                                 if connection.isVideoMirroringSupported {
-                                    connection.isVideoMirrored = true
+                                    connection.isVideoMirrored = self.isVideoMirrored
                                 }
 //                                #if compiler(<5.8.1)
                                     if #available(macOS 14.0, *), connection.isVideoRotationAngleSupported(self.orientation){
@@ -519,7 +518,7 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                             self.captureSession.addOutput(videoOutput)
                             for connection in videoOutput.connections {
                                 if connection.isVideoMirroringSupported {
-                                    connection.isVideoMirrored = true
+                                    connection.isVideoMirrored = self.isVideoMirrored
                                 }
 //                                #if compiler(<5.8.1)
                                 if #available(macOS 14.0, *),  connection.isVideoRotationAngleSupported(self.orientation){
@@ -837,7 +836,9 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                                             if self.isRecording && videoWriter.status == .writing {
                                                 self.stopRecording { callbackResult in
                                                     if let outputChannel = self.outputChannel {
-                                                        outputChannel.invokeMethod("onVideoRecordingFinished", arguments: callbackResult)
+                                                        DispatchQueue.main.async {
+                                                            outputChannel.invokeMethod("onVideoRecordingFinished", arguments: callbackResult)
+                                                        }
                                                     }
                                                 }
                                             }